requirements.txt
llama-cpp-python
PyQt5==5.15.11
python-docx==1.2.0
pdfplumber==0.11.8
requests==2.32.5
reportlab==4.4.5
PyInstaller==6.17.0
tqdm==4.67.1
lxml==6.0.2
pypdfium2==5.1.0
packaging==25.0
altgraph==0.17.5
pefile==2024.8.26
pywin32-ctypes==0.2.3
certifi
